+++ changelog.md
@@ -15,6 +15,10 @@
 
 - [#4691](https://github.com/ignite/cli/pull/4691), [#4706](https://github.com/ignite/cli/pull/4706), [#4725](https://github.com/ignite/cli/pull/4725) Fix ts-client query template and solely Go template for `ts-client` generation.
 
+### Features
+
+- [#4728](https://github.com/ignite/cli/pull/4728) Ask before overwriting a `xgenny` file.
+
 ## [`v29.0.0`](https://github.com/ignite/cli/releases/tag/v29.0.0)
 
 ### Features